% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betasigmoid.R
\name{dbetasigmoid}
\alias{dbetasigmoid}
\alias{rbetasigmoid}
\alias{pbetasigmoid}
\alias{qbetasigmoid}
\title{Beta-Sigmoid probability distribution}
\usage{
dbetasigmoid(x, a = 1, b = 1, infl = 0, scale = 1, log = FALSE)

rbetasigmoid(n, a = 1, b = 1, infl = 0, scale = 1)

pbetasigmoid(q, a = 1, b = 1, infl = 0, scale = 1,
  lower.tail = TRUE, log.p = FALSE)

qbetasigmoid(p, a = 1, b = 1, infl = 0, scale = 1,
  lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{x, q}{vector of quantiles}

\item{a, b}{non-negative parameters of the Beta-Sigmoid distribution which can be understood as number of success and failures.
They correspond to the `shape1` and `shape2` parameters of the Beta distribution. Default: 1}

\item{infl}{position of the inflection point of the sigmoid. Default: 0}

\item{scale}{positive number that describes how broad the sigmoid is. Default: 1}

\item{log, log.p}{logical; if `TRUE`, probabilities are given as log(p)}

\item{n}{number of observations}

\item{lower.tail}{logical; if `TRUE` (default), probabilities are `P[X < x]`, otherwise `P[X > x]`
\deqn{x \sim \mathrm{BetaSigmoid}(\alpha, \beta)}{x ~ BetaSigmoid(a, b, i, s)}}

\item{p}{vector of probabilities}
}
\description{
The distribution is obtained as a transformation of the Beta distribution with a sigmoid function. It
is useful for modelling the distribution of a mean if all that is known is a number of success and
failures and the sigmoid distribution that assignes the success and failures:
\deqn{m \sim \mathrm{Bernoulli}(\mathrm{sigmoid(\mu)})}{b ~ Bernoulli(sigmoid(µ))}
where a is number of times that m is 1 and b is the number of times it is 0.
}
\details{
The probability density function of the Beta-Sigmoid distribution is
\deqn{f_{BS_{\alpha, \beta, i, s}}(x) = \frac{\Gamma(\alpha + \beta)}{\Gamma(\alpha) \Gamma(\beta)}
       \Big(\frac{1}{1+e^{-x}}\Big)^\alpha \Big(1  - \frac{1}{1+e^{-(x - i) / s}}\Big)^\beta }
}
\examples{
dbetasigmoid(4, a=3, b=1, infl=3.5, scale=3)
pbetasigmoid(-60, a=0.3, b=10, infl=3.5, scale=5)
qbetasigmoid(0.8, a=3, b=3, infl=0, scale=0.5)
rbetasigmoid(3, a=19, b=39)
}
