% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betasigmoid.R
\name{logit}
\alias{logit}
\alias{sigmoid}
\title{Logit and Sigmoid function}
\usage{
logit(x, infl = 0, scale = 1, oneminus = FALSE)

sigmoid(x, infl = 0, scale = 1, log = FALSE, oneminus = FALSE)
}
\arguments{
\item{x}{numerical vectors}

\item{infl}{position of the inflection point of the sigmoid}

\item{scale}{the width of the sigmoid}
}
\description{
Logit and Sigmoid function
}
